more improvements to send all the messages in one go

+++ src/main/java/person/prashant/qpid/publisher/OurMessageProducer.java
@@ -4,6 +4,6 @@ package person.prashant.qpid.publisher;
 import javax.jms.JMSException;
 
 public interface OurMessageProducer {
-    void publishFixedNumberOfMessages(int messageCount) throws JMSException;
+    void publishFixedNumberOfMessages(long messageCount) throws JMSException;
     void keepPublishingMessagesAfterProvidedMilliSeconds(int timeGapInMillis) throws JMSException;
 }
